Модульность файлов
прописаны импорт и экспорт
