+++ index.html
@@ -235,6 +235,5 @@
     </template>
 
 </body>
-<script src="js/main.js"></script>
-<script src="js/functions.js"></script>
+<script type="module" src="js/main.js"></script>
 </html>
