+++ js/data.js
@@ -1,0 +1,22 @@
+// Массив имен
+const NAMES = ['Иван', 'Никита', 'Максим', 'Леонид', 'Кирил', 'Алексей', 'Денис', 'Настя', 'Марина', 'Ника', 'Даша', 'Майя', 'Эля', 'Диана'];
+
+// Массив сообщений
+const MESSAGES = ['Всё отлично!', 'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.', 'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+// количество лайков
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+
+// количество коментов
+const MIN_COMMENTS = 0;
+const MAX_COMMENTS = 30;
+
+//число обьектов
+const OBJ_NUMBER = 25;
+
+export {NAMES, MESSAGES, MIN_LIKES, MAX_LIKES, MIN_COMMENTS, MAX_COMMENTS, OBJ_NUMBER};
