+++ js/functions.js
@@ -1,10 +1,5 @@
 // 1я функция
 
-// function getStringLength(string, maxSymbols) {
-//   const stringLength = string.length;
-//   return stringLength <= maxSymbols;
-// }
-
 const getStringLength = (string, maxSymbols) => (string.length <= maxSymbols);
 
 console.log(getStringLength('проверяемая строка', 20));
@@ -12,25 +7,6 @@ console.log(getStringLength('проверяемая строка', 18));
 console.log(getStringLength('проверяемая строка', 10));
 
 // 2я функция
-
-// function isPalindrome (string) {
-//   const stereotypedString = string.replaceAll(' ', '').toLowerCase();
-//   let mirrorString = '';
-//
-//   for (let i = stereotypedString.length; i; i--) {
-//     mirrorString += stereotypedString[i - 1];
-//   }
-//   return mirrorString === stereotypedString;
-// }
-
-/* или
-let isPalindrome = (string) => {
-  let stereotypedString = string.replaceAll(' ', '').toLowerCase();
-  let mirrorString = stereotypedString.split('').reverse().join('');
-  return mirrorString === stereotypedString;
-}; */
-
-/* или */
 
 const isPalindromeByArrayMethod = (string) => {
   const stereotypedString = string.replaceAll(' ', '').toLowerCase();
@@ -65,3 +41,5 @@ console.log(extractNumbers('а я томат'));
 console.log(extractNumbers(2023));
 console.log(extractNumbers(-1));
 console.log(extractNumbers(1.5));
+
+export {getStringLength, extractNumbers, isPalindromeByArrayMethod};
